Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/execution-timing/106-import-xhtml-expected.txt

Test

FAIL Stylesheet in XHTML HEAD with @import blocking scripts assert_equals: expected "fixed" but got "static"
